Remove reliance on Ruby
First, because there are CVEs in some of our ruby dependencies that
would require us to upgrade.

Second, because nobody in our team has decent experience in Ruby,
so it's kind of nuts to use it in such a critical piece of our
infrastructure.

And finally, because we already have alternative shell scripts to do
what the ruby scripts do, and we've already used those scripts in
several releases, for both Hibernate Search and Hibernate Validator.

The only project that hasn't been tested with the shell scripts is OGM,
and, well, it's been [retired](https://hibernate.org/others/#hibernate-ogm).
So, we'll see about that when/if we release OGM again.